test(github): split client tests into http_tests + classify_tests
github/tests.rs (503) split into sibling http_tests.rs and classify_tests.rs,
both <=400 lines, matching the brave/client and fetch/ssrf layout. Production
github.rs is unchanged apart from the mod declaration; all 449 tests pass.
